requirements.txt
Flask
Flask-SocketIO==5.2.0
simple-websocket
kombu
importlib-metadata==4.13.0
python-decouple
